fix: 首页/动态的recyclerview适配了圆屏（写了我一个晚上的曲线列表

+++ app/src/main/java/cn/spacexc/wearbili/dataclass/dynamic/Card.kt
@@ -31,7 +31,7 @@ data class Card(
     val stype: Int,
     val timestamp: Long,
     val type: Int,
-    val uid: Int,
+    val uid: Long,
     val uid_type: Int,
     val user_profile: UserProfile,
     val view: Int
